Move DaysFilter to common
