+++ src/components/common/DaysFilter/DaysFilter.stories.tsx
@@ -4,7 +4,7 @@ import { DaysFilter } from ".";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
 const meta: Meta<typeof DaysFilter> = {
-  title: "Errors/GlobalErrorsList/DaysFilter",
+  title: "common/DaysFilter",
   component: DaysFilter,
   parameters: {
     // More on how to position stories at: https://storybook.js.org/docs/react/configure/story-layout
