+++ src/components/common/DaysFilter/index.tsx
@@ -1,16 +1,16 @@
 import type { ChangeEvent } from "react";
 import { useCallback, useEffect, useMemo, useState } from "react";
-import { usePrevious } from "../../../../hooks/usePrevious";
-import { isUndefined } from "../../../../typeGuards/isUndefined";
-import { sendUserActionTrackingEvent } from "../../../../utils/actions/sendUserActionTrackingEvent";
-import { addPrefix } from "../../../../utils/addPrefix";
-import { formatUnit } from "../../../../utils/formatUnit";
-import { CalendarIcon } from "../../../common/icons/12px/CalendarIcon";
-import { MinusIcon } from "../../../common/icons/MinusIcon";
-import { PlusIcon } from "../../../common/icons/PlusIcon";
-import { NewPopover } from "../../../common/NewPopover";
-import { NewIconButton } from "../../../common/v3/NewIconButton";
-import { MenuList } from "../../../Navigation/common/MenuList";
+import { usePrevious } from "../../../hooks/usePrevious";
+import { isUndefined } from "../../../typeGuards/isUndefined";
+import { sendUserActionTrackingEvent } from "../../../utils/actions/sendUserActionTrackingEvent";
+import { addPrefix } from "../../../utils/addPrefix";
+import { formatUnit } from "../../../utils/formatUnit";
+import { MenuList } from "../../Navigation/common/MenuList";
+import { CalendarIcon } from "../icons/12px/CalendarIcon";
+import { MinusIcon } from "../icons/MinusIcon";
+import { PlusIcon } from "../icons/PlusIcon";
+import { NewPopover } from "../NewPopover";
+import { NewIconButton } from "../v3/NewIconButton";
 import * as s from "./styles";
 import { trackingEvents } from "./tracking";
 import type { DaysFilterProps } from "./types";
