+++ src/featureFlags.ts
@@ -34,7 +34,7 @@ export const featureFlagMinBackendVersions: Record<FeatureFlag, string> = {
   [FeatureFlag.IsUserIdEnabled]: "0.3.212",
   [FeatureFlag.IsIssuesCriticalityLevelsFilterEnabled]: "0.3.214",
   [FeatureFlag.IsNewBehaviorForMetricsTimeModesEnabled]: "0.3.274",
-  [FeatureFlag.IsIssuesLastDaysFilterEnabled]: "0.3.274"
+  [FeatureFlag.IsIssuesLastDaysFilterEnabled]: "0.3.277-alpha.3"
 };
 
 export const getFeatureFlagValue = (
